// frontend/hooks/models/playlistMemberDTO.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface PlaylistMemberDTO {
  userId?: number;
  username?: string;
  displayName?: string;
  avatarUrl?: string;
  owner?: boolean;
  canRead?: boolean;
  canWrite?: boolean;
  canDelete?: boolean;
  joinedAt?: string;
}